Moved occations folder to tabs folder
